pyproject.toml
[project]
name = "smp-agents"
version = "0.1.0"
description = "Skills Marketplace ADK agents: Skill Advisor, Bundle Validator, KG Q&A, Playground, and Skill Builder."
readme = "README.md"
license = {text = "Apache License 2.0"}
requires-python = ">=3.11"

dependencies = [
    "google-adk[a2a]>=1.25.0,<2.0.0",
    "litellm>=1.30.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "neo4j>=5.0.0,<6.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "google-adk[eval]>=1.25.0",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "ruff>=0.8.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = [
    "shared",
    "agents/skill_advisor",
    "agents/bundle_validator",
    "agents/kg_qa",
    "agents/playground",
    "agents/skill_builder",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests with mocked dependencies",
    "integration: Integration tests requiring live Neo4j and LLM",
    "e2e: End-to-end tests against deployed agents on OpenShift",
    "eval: ADK agent evaluation tests requiring live LLM and Neo4j",
]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["shared", "agents"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
